third_party/WebKit/LayoutTests/virtual/without-smil/svg/animations/exposed/interface.html

<!DOCTYPE html>
<html>
<head>
  <script src="../../../../../resources/js-test.js"></script>
</head>
<body>
  <svg id="svg" xmlns="http://www.w3.org/2000/svg">
    <rect width="100" height="200">
      <set id="animation" attributeName="width" to="300"/>
    </rect>
  </svg>

  <script>
    'use strict';

    var svg = document.getElementById('svg');
    var animation = document.getElementById('animation');

    shouldBeUndefined("svg.animationsPaused");
    shouldBeUndefined("svg.getCurrentTime");
    shouldBeUndefined("svg.pauseAnimations");
    shouldBeUndefined("svg.setCurrentTime");
    shouldBeUndefined("svg.unpauseAnimations");

    shouldBeUndefined("animation.beginElement");
    shouldBeUndefined("animation.beginElementAt");
    shouldBeUndefined("animation.endElement");
    shouldBeUndefined("animation.endElementAt");
    shouldBeUndefined("animation.getCurrentTime");
    shouldBeUndefined("animation.getSimpleDuration");
    shouldBeUndefined("animation.getStartTime");
    shouldBeUndefined("animation.onbegin");
    shouldBeUndefined("animation.onend");
    shouldBeUndefined("animation.onrepeat");
    shouldBeUndefined("animation.targetElement");

    if (window.testRunner)
        testRunner.dumpAsText();
  </script>

  <p id="description"></p>
  <div id="console"></div>
</body>
</html>
