third_party/WebKit/LayoutTests/virtual/without-smil/svg/animations/exposed/interface-expected.txt

PASS svg.animationsPaused is undefined.
PASS svg.getCurrentTime is undefined.
PASS svg.pauseAnimations is undefined.
PASS svg.setCurrentTime is undefined.
PASS svg.unpauseAnimations is undefined.
PASS animation.beginElement is undefined.
PASS animation.beginElementAt is undefined.
PASS animation.endElement is undefined.
PASS animation.endElementAt is undefined.
PASS animation.getCurrentTime is undefined.
PASS animation.getSimpleDuration is undefined.
PASS animation.getStartTime is undefined.
PASS animation.onbegin is undefined.
PASS animation.onend is undefined.
PASS animation.onrepeat is undefined.
PASS animation.targetElement is undefined.
PASS successfullyParsed is true

TEST COMPLETE
